+++ Editor/StateGraph/VisualElements/StateGraphView.cs
@@ -23,6 +23,11 @@ namespace Nonatomic.VSM2.Editor.StateGraph
 		{
 			MakeTitleBar();
 		}
+
+		public void OnFocus()
+		{
+			Debug.Log("OnFocus");
+		}
 		
 		protected override void MakeStateManager(string id)
 		{
@@ -37,6 +42,7 @@ namespace Nonatomic.VSM2.Editor.StateGraph
 				HandleRecenter();
 			}
 			
+			Debug.Log($"PopulateGraph: {model}");
 			base.PopulateGraph(model);
 
 			_titleBar.SetTitle(model.name);