+++ CHANGELOG.md
@@ -1,4 +1,11 @@
 # Change Log
 
-## [0.0.0] - Feb 03, 2024
-- First release of the Visual State Machine V2 package
+## [0.0.1-beta] - Feb 11, 2024
+- Fixed issue with the `StateMachineEditorWindow` not loading the last model when Unity opened.
+- Added custom inspector for the `StateMachineController` to show the create and open state machines.
+
+## [0.0.0-beta] - Feb 08, 2024
+- First tagged release
+
+## [0.0.0-beta] - Feb 03, 2024
+- First release of the Visual State Machine V2 package[package.json](package.json)[package.json](package.json)[package.json](package.json)